Разбил TokenReader на несколько классов
