+++ cs/MarkdownTests/MarkdownTests.csproj
@@ -33,8 +33,12 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="MarkdownTests.cs" />
-    <Compile Include="MarkdownTestsData.cs" />
+    <Compile Include="MdNormalizerTests\MdNormalizerTests.cs" />
+    <Compile Include="MdNormalizerTests\MdNormalizerTestsData.cs" />
+    <Compile Include="MdRendererTests\MarkdownTests.cs" />
+    <Compile Include="MdRendererTests\MarkdownTestsData.cs" />
+    <Compile Include="TokenReaderTests\TokenReaderTests.cs" />
+    <Compile Include="TokenReaderTests\TokenReaderTestsData.cs" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="FluentAssertions, Version=5.9.0.0, Culture=neutral, PublicKeyToken=33f2691a05b67b6a">
